requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Werkzeug==3.0.1

# Security
bcrypt==4.1.2

# Utilities
tzdata==2024.1
requests==2.31.0
beautifulsoup4==4.12.3

# Production WSGI server
gunicorn==21.2.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-flask==1.3.0
